CSID-460: Adding cloud provider clients validation on replicator demos
I am also fixing the GKE demo, as it had the same shortcoming.
This would cause the demo to fail if the az client is not installed or if the resource_group variable is not defined,  but without a proper error message as it is not checked.
Similar issue for GKE
@rspurgeon