Use return value from ccloud call instead of env variable (#1024)
n the ccloud::destroy_ccloud_stack() function, the check to determine if a ksqlDB needs to be deleted is based on environment variable KSQLDB_ENDPOINT. But there could be cases where the variable is not set or gets un-set. This situation causes an error and the cluster can't be deleted as the ksqkDB application is still running.

Since the function asks the user to pass in the SERVICE_ACCOUNT_ID and with that parameter we set the KSQLDB_NAME variable. We can then deterministically figure out if there is a ksqlDB application by running the ccloud command to list ksql applications and grep for the name.

This way the delete function isn't relying on any environment parameters and a user can open up a new terminal window and still successfully remove all CCloud resources.